requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
Werkzeug==3.1.3
python-dotenv==1.0.0
requests==2.31.0

